+++ pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>de.craftlancer.imagemaps</groupId>
   <artifactId>ImageMaps</artifactId>
-  <version>0.5.0</version>
+  <version>0.5.1</version>
   <name>ImageMaps</name>
   <description>Draw Images on maps!</description>
   <properties>
