+++ src/main/java/de/craftlancer/imagemaps/ImageMaps.java
@@ -1,14 +1,17 @@
 package de.craftlancer.imagemaps;
 
-import de.craftlancer.imagemaps.services.MapService;
-import de.craftlancer.imagemaps.services.impl.MapServiceImpl;
-import org.bukkit.event.Listener;
+import de.craftlancer.imagemaps.command.ImageMapCommand;
+import de.craftlancer.imagemaps.event.PlayerInteractions;
+import de.craftlancer.imagemaps.service.MapService;
+import de.craftlancer.imagemaps.service.impl.MapServiceImpl;
+import de.craftlancer.imagemaps.task.FastSendTask;
+import de.craftlancer.imagemaps.task.ImageDownloadCompleteNotifier;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.io.File;
 import java.util.logging.Level;
 
-public class ImageMaps extends JavaPlugin implements Listener {
+public class ImageMaps extends JavaPlugin {
     public static final int MAP_WIDTH = 128;
     public static final int MAP_HEIGHT = 128;
     public static final String IMAGES_DIR = "images";
@@ -31,7 +34,7 @@ public class ImageMaps extends JavaPlugin implements Listener {
 
         mapService.loadMaps();
         getCommand("imagemap").setExecutor(new ImageMapCommand(this));
-        getServer().getPluginManager().registerEvents(this, this);
+        getServer().getPluginManager().registerEvents(new PlayerInteractions(), this);
         sendTask = new FastSendTask(this, mapsPerSend);
         getServer().getPluginManager().registerEvents(sendTask, this);
         sendTask.runTaskTimer(this, sendPerTicks, sendPerTicks);
