+++ src/main/java/de/craftlancer/imagemaps/model/ImageMap.java
@@ -1,4 +1,4 @@
-package de.craftlancer.imagemaps;
+package de.craftlancer.imagemaps.model;
 
 public class ImageMap
 {
