+++ src/main/java/de/craftlancer/imagemaps/model/PlacingCacheEntry.java
@@ -1,0 +1,29 @@
+package de.craftlancer.imagemaps.model;
+
+public class PlacingCacheEntry
+{
+    private String image;
+    private boolean fastsend;
+    private double scale;
+    
+    public PlacingCacheEntry(String image, boolean fastsend, double scale)
+    {
+        this.image = image;
+        this.fastsend = fastsend;
+        this.scale = scale;
+    }
+    
+    public String getImage()
+    {
+        return image;
+    }
+    
+    public boolean isFastSend()
+    {
+        return fastsend;
+    }
+    
+    public double getScale() {
+        return scale;
+    }
+}
